moved functions to module file
